server-client/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "homelab-monitor"
version = "1.0.0"
description = "Server monitoring client for HomeLab Monitor display"
readme = "../README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "psutil>=5.9.0",
    "pyserial>=3.5",
]

[project.scripts]
homelab-monitor = "src.monitor:main"

[project.urls]
Homepage = "https://github.com/yourusername/homelab-monitor"
Repository = "https://github.com/yourusername/homelab-monitor.git"
Issues = "https://github.com/yourusername/homelab-monitor/issues"
